+++ src/shift_left/tests/it/debug_it.py
@@ -32,9 +32,10 @@ class TestDebugIntegrationTests(unittest.TestCase):
         #result = runner.invoke(app, ['table', 'migrate', 'dim_training_course', os.getenv('SRC_FOLDER','.') + '/dimensions/qx/dim_training_course.sql', os.getenv('STAGING')])
         #result = runner.invoke(app, ['table', 'init-unit-tests', 'aqem_fct_step_role_assignee_relation'])
         #result = runner.invoke(app, ['table', 'build-inventory'])
-        #result = runner.invoke(app, ['pipeline', 'build-metadata', os.getenv('PIPELINES') + '/sources/qx/audit_trail/sql-scripts/dml.src_qx_audit_trail.sql'])
+        #result = runner.invoke(app, ['pipeline', 'build-metadata', os.getenv('PIPELINES') + '/dimensions/aqem/dim_event_action_item/sql-scripts/dml.aqem_dim_event_action_item.sql'])
         #result = runner.invoke(app, ['table', 'run-unit-tests', 'aqem_dim_event_element', '--test-case-name', 'test_aqem_dim_event_element_1'])
-        result = runner.invoke(app, ['pipeline', 'build-execution-plan', '--product-name', 'aqem'])
+        result = runner.invoke(app, ['pipeline', 'deploy', '--product-name', 'aqem', '--max-thread' , 10, '--pool-creation'])
+        # result = runner.invoke(app, ['pipeline', 'undeploy', '--product-name', 'aqem', '--no-ack'])
         print(result.stdout)