+++ src/shift_left/tests/ut/core/test_execution_plan.py
@@ -34,6 +34,7 @@ class TestExecutionPlan(BaseUT):
           \          \
     src_a -> a        \
     src_b -> b ------>  c -> e
+    Added more nodes to test cross product.
     """
     
     TEST_COMPUTE_POOL_ID_1 = "lfcp-121"
@@ -708,7 +709,6 @@ class TestExecutionPlan(BaseUT):
                                 "dev-usw2-p2-dml-b"]:  
                 return self._create_mock_get_statement_info(name=statement_name,status_phase="RUNNING")
             else:
-                
                 return self._create_mock_get_statement_info(name=statement_name, status_phase="UNKNOWN") 
         
         mock_get_status.side_effect = mock_statement
@@ -723,7 +723,8 @@ class TestExecutionPlan(BaseUT):
             compute_pool_id=self.TEST_COMPUTE_POOL_ID_1,
             execute_plan=False,
             may_start_descendants=True,
-            force_ancestors=False
+            force_ancestors=False,
+            pool_creation=False
         )
         print(f"{summary}\n")
         assert len(report.tables) == 13